\name{compute.Stats}
\alias{compute.Stats}

\title{
compute Stats from scores and null distribution
}
\description{

}
\usage{
compute.Stats(null.distrib, final.scores, nb.genes, nperms)
}

\arguments{
  \item{null.distrib}{

}
  \item{final.scores}{

}
  \item{nb.genes}{

}
  \item{nperms}{

}
}
\details{

}
\value{

}
\references{

}
\author{

}
\note{

}


\seealso{

}
\examples{
}
\keyword{ ~kwd1 }
