\name{construct.df.scores}
\alias{construct.df.scores}

\title{
construct data frame of scores for paired comparisons
}
\description{

}
\usage{
construct.df.scores(lMs, lOs, lIsPaireds, pairs)
}

\arguments{
  \item{lMs}{

}
  \item{lOs}{

}
  \item{lIsPaireds}{

}
  \item{pairs}{

}
}
\details{

}
\value{

}
\references{

}
\author{

}
\note{

}


\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
