\name{correctlocalQval}
\alias{correctlocalQval}

\title{
from local to global q-values
}
\description{
assures that q-values are always increasing for increasing scores/p-values
}
\usage{
correctlocalQval(qvals)
}

\arguments{
  \item{qvals}{
    vector of ordered q-values (from lowest to highest p-values or from
  highest to lowest score)
}
}
\details{

}
\value{
}
\references{

}
\author{
maxime.tarabichi@ulb.ac.be
}
\note{

}


\seealso{

}
\examples{

}
\keyword{ ~kwd1 }
