\name{find.pairs}
\alias{find.pairs}

\title{
finds all possible pairs of sample to compare
}
\description{

}
\usage{
find.pairs(lOs)
}

\arguments{
  \item{lOs}{

}
}
\details{

}
\value{

}
\references{

}
\author{
maxime.tarabichi@ulb.ac.be
}
\note{

}


\seealso{

}
\examples{

}
\keyword{ ~kwd1 }
