\name{giveRPsymetric.divideMatrix}
\alias{giveRPsymetric.divideMatrix}

\title{
gives RP scores optimizing memory consumption
}
\description{

}
\usage{
giveRPsymetric.divideMatrix(lMs, lOs, doPrint = FALSE, size = 64, mc.cores)
}

\arguments{
  \item{lMs}{

}
  \item{lOs}{

}
  \item{doPrint}{

}
  \item{size}{

}
  \item{mc.cores}{

}
}
\details{

}
\value{

}
\references{

}
\author{

}
\note{

}


\seealso{

}
\examples{

}
\keyword{ ~kwd1 }
