\name{giveRPsymmetric}
\alias{giveRPsymmetric}

\title{
gives RP scores 
}
\description{

}
\usage{
giveRPsymmetric(lMs, lOs, pairs, doPrint = FALSE)
}

\arguments{
  \item{lMs}{
list of matrices in the meta-analysis. A list containing one matrix can
  be passed.
}
  \item{lOs}{
list of 2-level factors corresponding to the matrices in lMs. Attention
  must be paid to the order of the levels in each factor. Ratios are
  always level1/level2. Ideally, each vector should heave the two same
  unordered levels.
}
  \item{pairs}{
    not used.
}
  \item{doPrint}{
    used internally by the package.
}
}
\details{

}
\value{

}
\references{

}
\author{
maxime.tarabichi@ulb.ac.be
}
\note{

}



\seealso{

}
\examples{

}
\keyword{ ~kwd1 }
