\name{RPlite-package}
\alias{RPlite-package}
\alias{RPlite}
\docType{package}
\title{
RankProduct methodology. Meta-analysis of multiple datasets. Memory usage optimized. Parallelized with dependency
on library multicore.
}
\description{

}
\details{
\tabular{ll}{
Package: \tab RPlite\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-15\cr
License: \tab What license is it under?\cr
}
}
\author{
please write to maxime.tarabichi@ulb.ac.be for any comments
}
\references{
Breitling et al. Rank Products: a simple, yet powerful, new method to
detect differentially regulated genes in replicated microarray experiments.
}
\keyword{ package }
\seealso{
GSEAlite
}
\examples{
see function RP.symetric.Cp and RPpaired
}
